All Games|Fixed: A_Lower action could result in an undefined state
Attempting to use the A_Lower action to forcibly lower the weapon
during e.g., the weapon ready state sequence, resulted in the game
entering an undefined state which stopped all weapon changes from
functioning normally.

When A_Lower is now used in such a manner, the weapon is indeed
lowered but it will remain so - no subsequent weapon change occurs
and no other weapon will be raised. The user is still allowed to
change weapon in this state.