Ded Reader|Added: State defs now support the Mods directive
The 'Mods' directive allows the properties of an existing definition
to be modified instead of creating a new one. This functionality can
be used to "patch" the original definitions, changing only some of
their properties.

Also fixed a bug in the handling of the 'Copy' directive which failed
to duplicate the command execution string (resulting in an attempt to
free the one string twice during shutdown/game change).