HacX|Fixed: Password game object was broken
Make use of the State 'Mods' directive to modify the original game
definitions instead of replacing them (for the purpose of tweaking
lighting properties such as "fullbright"). The DEHACKED patch reader
looks up definitions by the ordinals used in the original executable
and as such, we can't replace them with later ones (load order).

Resolves a regression introduced in build 402.