Doom: Check floating monster position after Z move
Floating monsters were not ensuring that the Z move leads
to a valid position. The monster was first moved on the XY
plane, but after floating a check was never made about
whether the mobj came into contact with something.

Note: In the original Doom all movement checks were 2D,
but Doomsday uses 3D checks.