Input|Bindings: Mark toggles triggered, check the flag for the Attack control
The Attack control is not an impulse control. This means that if the key is
not held down when a sharp tick occurs, it is not considered to be active
even though it has been pressed and released since the previous sharp
tick.

The trigger flags are drawn in the key state visualization as purple marks.