Hexen|Fixed: Errors in player-on-mobj logic
Hexen's on-mobj tests use a different logic than the other games.
It was not taking into account the situation where the player was
standing on a mobj and contacting another mobj on which it was
also possible to stand on. This would result in CheckPosition from
skipping line checks altogether.

View bobbing was not working when standing on a mobj because
the player mobj's floorZ was not adjusted to the top of the onMobj;
the view height calculation considered the player airborne when
standing on a thing.