Ded Reader|Added: Thing defs now support the Mods directive
The 'Mods' directive allows the properties of an existing definition
to be modified instead of creating a new one. This functionality can
be used to "patch" the original definitions, changing only some of
their properties.

Also added the alternative name "Mobj" for this definition type, for
consistency with the rest of the public API.