Memory Zone: Disabled fast zmalloc mode
It seems that the fast mode results in the allocation of excessive
memory volumes.

Todo for later: We need to revise the memory zone for more
efficient utilization of the allocated space. Possibilities: purge level
specific volumes, do allocs with system malloc(), reduce block
header size (currently 64 bytes in a 64-bit build…).