File Sys|Fixed: -iwad failed locating iwads in root of Doomsday install dir
FileDirectory by default will attempt to remove the engine's install
dir when inserting paths, i.e., it tries to make all paths relative
when it can so as to compress the data (there is no need to represent
the base multiple times). Consequently when presented with a path
exactly equal to the install dir; the resultant relative path became
a zero-length string, which was then not added to the directory.

FileDirectory can now be initialized in one of two distinct modes:

A) Relative mode
Both relative and absolute paths may be inserted into the directory.
The base path is defined during instantiation. An attempt is made to
transform all paths into relative ones upon being added into the
directory.

B) Absolute mode
Only absolute paths may be inserted into the directory. All paths are
represented using their fully qualified paths (i.e., the base is not
removed from paths).

In relative mode, FileDirectory will now insert a new node into it's
internal PathDirectory to represent the symbolic "./" directory when
it is first added. This symbolic node is used to represent the base
directory which it will by default remove in the process of making
paths relative.