Memory Zone|Debug|Refactor: Memory debug visualization and tweaks
New memory volumes are added to the end of the volume list so that the
earlier ones still get utilized after a new one is added. Helps prevent
the addition of more and more volumes.

Refactored Z_Malloc()'s rover movement for clarity.

The debug visualization is enabled with the -zonedebug option.