Memory Zone: Observe when a volume fills up
A counter of allocated bytes is kept for each volume. When the
volume gets too full according to the counter, the volume is
skipped when doing new mallocs.

Fixes a performance problem where a full volume would cause
a new malloc to require scanning through all the blocks to see
if any free space remains.