API: Use LineDef (and not Seg) with P_SetPolyobjCallback()
The original game would use the seg angles for the purpose of testing
collisions with mobjs. This rather defied the design of the idtech 1
engine which otherwise used linedefs everywhere else for this purpose.
This exposing of a polyobj's segs to the playsim is unnecessary.

Polyobjs now use linedef angles for collision purposes.