"unload" should be called at Sys_Shutdown time
Irrespective of whether the engine is shutting down normally or in
response to a fatal error, any currently loaded game plugin must be
unloaded before we can terminate.

Not calling "unload" the game configs and control bindings will not
be saved.

In the case of a fatal error the global appShutdown variable must be
set before the unload, else the engine will re-enter ringzero.