Audio|Refactor: Allow loading multiple audio plugins
Refactored the audio plugin loader to allow loading more than one
plugin at a time. The loading and initialization of the audio
plugins was moved to the new AudioDriver module.

The loading was also made platform-independent through the use
of the Library class (which presently lacks a Windows implementation).

The AudioDriver module chooses which SFX, Music, and CD interface
to use from which loaded plugin.