Refactor: Moved sector and sidedef origins to members of Sector and SideDef
Also fixed a bug with polyobjects in that the sidedef origins were
not being updated after a polyobject has moved (resulting in the
distance based delta queueing using the old origins when subsequent
changes to those sidedefs occur).

Todo for later: These origin points can be trivially calculated so
is there really a need to cache them?