qmake|Snowberry: Launcher script and .desktop file respect config.pri
The launch-doomsday script and the .desktop file for starting the script
are now generated by qmake so that their contents reflect the directories
configured in config.pri. (patch by vvv1)

Added the configuration variable PYTHON for manually specifying which
Python interpreter to use. The default (not specified) assumes that
Python can be found on the system path.