Fixed|GCC: Illegal and accidental dynamic linkage from plugin to engine
The WadMapConverter plugin was accidentally linked to a symbol
in the engine, causing the plugin to disrupt internal engine state.
Added a macro for naming plugin globals in such a way that they
won't clash with engine globals.

This won't be a problem any more after we apply the libdeng
refactoring...