Improved screen grabbing; default screenshot format is now .png
Added routines for grabbing the contents of a Canvas in various
ways. One can create an image_t, QImage or an OpenGL texture
out of the contents.

Improved performance of busy mode screen grab by making it
straight into a texture.

Screen captures are now saved as .png instead of uncompressed
Targa. This should make life easier for people who take
screenshots.

Todo: Test what happens on old graphics cards that don't
support non-power-of-two textures. I'm not sure how QGLContext
handles the QImage->texture creation in that case (either
scale or sub-image). If the non-Po2 extension is not listed by
OpenGL, we can always pre-scale the QImage to a Po2 size.