Win32: Use DirectInput for mouse
It seems that the existing DirectInput code is the more robust way to get
mouse delta events in Windows. Refactored the low-level mouse code by adding
a light-weight mouseinterface_t that points to the actual mouse functions.

Added a separate source file for DirectInput, and moved the Win32 mouse code
to mouse_win32.c.