BSP Builder|Refactor: Segregated build completion tasks (SoC)
The BspBuilder's destructor will now delete all biproduct objects
produced during the course of ::build().

The actual BSP data objects (BspNode, BspLeaf and HEdge) are not
destroyed as it is assumed that MPE_SaveBSP will have been called
to obtain ownership of it.

Todo: BspBuilder should fully clean up after itself regardless of
the outcome or whether anyone decides to take ownership of all the
data objects it produces.