Added: Variables "vid-fsaa" and "vid-vsync"
FSAA and vsync can be toggled at runtime with the new vid-fsaa and
vid-vsync variables. The command line options -nofsaa and -novsync
can be used to permanently disable FSAA and vsync, respectively.

Also, when the mouse is trapped in a window, all GL updates will be
timed manually to match the requested maximum refresh rate. This
will only have an effect if refresh-rate-maximum is less than the vsync
rate.