Resources|Fixed: Locating all IWAD alternatives
If a game specified multiple names for an IWAD, the resource
locator would find one of them and only then check if it contained
the required lumps. However, some games may use the same
file name with different lump conditions (full Hexen vs. Hexen
demo). In this case, the locator failed to find the demo IWAD
because it only checked the file it found first -- the full IWAD.

The game resource validator was modified to check each
alternative name separately. The first one to fulfill all conditions
is selected for use.

Todo for later: A more elegant solution would be to move the
resource validation logic into AbstractResource.