BspLeaf: Removed array of fvertex_t allocated for each leaf
As we are presently copying the vertex coordinates into a write
buffer before adding them to the render lists there is no point
duplicating the vertex pointers. We can easily derive the order
and populate the vertex write buffer directly by navigating the
ring of HEdges linked to BspLeaf.

We now determine the HEdge to use as a base and record that in
BspLeaf instead. The rest of the order can be determined from
the base and the arguments used at geometry write time.

Todo for later: Cache final vertex coordinates for all geometry
sets (all planes and both sky caps) in BspLeaf.