Refactor: Renamed sidedefsection_t as SideDefSection and moved to public API
Note that the indices of the sections have also been changed to match
those used by the games when serializing material change thinkers to
saved games. The order of which does not matter as far as the engine
is concerned.

Also added an API helper macro for converting a SideDefSection to the
equivalent DMU property modifier flag.