libdeng2: Ported cfparser.py from Snowberry (Info file parser)
The Snowberry cfparser.py is the parser that Snowberry uses for all its
configuration, profile, and add-on metadata Info files. Having this in libdeng2
will enable us to read these files natively in Doomsday.

Todo for later: Add tests for the parser.
Todo for later: Take advantage of the scriptsys parsing classes where applicable.