Hacx|Fixed: Added space character patch for the BIGFN font
By default the font renderer will complete all missing font character
geometry by populating the dimensions with an average of the present
characters.

As Hacx is missing a patch for the space character, this results in a
much larger gap between words when compared to the original patches.

This missing patch has been added to jdoom.pk3

Todo for later: Do not require a patch graphic for this. Instead we
should extend the Font definition to allow overrides for character
geometry to be specified there.