Map Renderer: Look for and fix single-sided linedefs missing a middle material
Fix missing materials on one-sided lines as well. The map author may
have overlooked texturing parts of their map and it is better to draw
*something* rather than HOM.

If we can't find a suitable replacement (unlikely, there is usually
a neighboring material) we'll instead choose the special "missing"
material as replacement.

Remember, the goal of the replacement logic is not to emulate how the
original game renderers handled the situation (for that we'll need to
use an entirely different approach).