Map Renderer|Fixed: Ignore the front ceiling when comparing sky-masked planes
Closer inspection revealed that doom.exe's column renderer does in
fact extend the lower wall section up to the height of the back
sector plane when both are sky-masked.

This fix addresses issues seen in ksutra.wad MAP01 and MAP05, which
are slight variations on the same construct.

Todo: There remains an issue with the Y texture coordinates on some
of the buildings in the start area of ksutra MAP01.