Win32|Fixed: Alt DDKEY not released when using alt+tab to switch tasks
On Windows the alt key-up event is eaten by the OS when using alt+tab
to switch tasks, prior to our Qt based event handler getting a chance
to process it.

This means it is necessary to monitor QEvent::ActivationChange events
in CanvasWindow and then manually submit a key-up event in response.

We must also track the is-down state of the alt key manually. Although
this is already implemented in the input module, this is too far away
for us to ignore repeat up events (due to cycling tasks and the manual
alt key-up submital).

I was unsure how well the input layer handles repeat state set events
so an additional state has been added to Canvas.

Todo for later: Check whether repeat key-up events are OK and if so,
remove the additional is-down state tracking in Canvas.