Input|Refactor: Perform mouse polling at a predictable point during the frame
The state of the mouse is now polled always at the same point in time
in relation to a frame update.

This ensures that the mouse information for each frame has been
acquired at the same point in time in relation to the update cycle
(providing the operating system provides high-precision mouse input).

Also cleans up the Canvas class slightly: trackMousePosition() is no
longer there and has been replaced with Qt-based mouse position
polling in mouse_qt.cpp.