Input: Sharp input applies to all input events (not just toggles)
Playsim uses sharp ticks to update view angles, so the mouse axis
position needs to be updated only at sharp ticks, if sharp events are
in use. Mouse filtering is now applied so that it takes the sharp
ticks into account.

The variable "input-sharp" replaces "input-toggle-sharp".

If mouse events were applied in fractional time, they would also
have to be applied to the view angles in fractional time.