Textures: Factor detail texture normalization into the contrast multiplier
Doomsday will by default normalize the luminance values of detail
textures so that they are balanced correctly. This means determining
the average luminance value (ideally 127) and using this to shift the
lighting values so that the average lies in the center of the range.
Values each side of the center point are then independently applified
to cover the whole 0..255 range. This luminance normalization will
remove any contrast factor that has been pre-applied in the texture
data.

Therefore this contrast factor must be re-applied when preparing the
detail texture for render.