Fine-tuned frame update timing
There are now two distinct modes that frame updates operate in:

1) Manual repaint mode. Offers maximum control over frame updates
and timing, with the expense of increased CPU usage. This mode is
used in fullscreen mode.

2) Automatic update mode. Frame updates and timing are controlled by Qt.
There is less CPU usage, but frame timing is not as accurate. This mode
is used in windowed mode.

Note: On X11, the automatic update mode is always used.

Todo: Verify that this works as expected on Windows.