ZipFile: Quick n' dirty fix to prevent a write past end of buffer
ZipFile_ReadLumpSection() was calling ZipFile_BufferLump() with the
user-given data buffer (which may be far smaller than the size of
the uncompressed lump data).

Allocate a new temporary buffer to hold the uncompressed data from
where we'll copy the subsection of that data the user requested.

Todo: Revert this commit once a better solution is implemented.