Map Renderer: Sort right wall divisions in ascending order
It is not necessary to sort the wall divisions for the right edge
in descending order as the two edges are handled independently in
any case (and we can inverse the order trivially by changing the
iteration direction).

Not sorting the divisions allows to potentially share the same set
for each side of an edge.