All Games: Added "game-save-last-loadonreborn" option
When enabled and the player asks to be reborn; attempt to automatically
load a game save state in the "last-used" slot.

Can be used in combination with "game-save-auto-loadonreborn" such
that the "last" slot is preferenced over the "auto" slot.

Note: Presently off by default.
Todo: When both options are set we should determine which save to load
by examining the serialized game state (newer auto saves should have
priority over older "manual" saves).