Hexen|Fixed: Crash when opening a swing door
This commit fixes a very old bug from 2003 that particularly affects
64-bit builds.

For doors of type PODOOR_SWING, the direction member is either 1
or -1. However, the erroneous PO_SetDestination() interpreted this
as a fine angle, which resulted in out-of-bounds memory access.

Here I've replaced calls to PO_SetDestination() with a polydoor
specific version that only operates on sliding doors (as swing doors
don't move, just rotate).