Fixed|All Games: Update Mobj "selector" health bits immediately on spawn
The "selector" is used when determining which skin to use for any
3D model defined for some mobj state. The health bits were not being
setup until P_MobjThinker was run for the first time, resulting in
the wrong skin being used during the screenwipe/transition.

Todo: There is likely a similar issue with the player - check that the
weapon bits are being set correctly.