Fixed|Hexen: Player colors in the Player Setup menu
In Hexen v1.0, there are only 9 translation tables available. When
loading the tables, they are mapped to indices 0-2, 6-8, and 13-15
in the engine's translation tables, as the engine assumes 7 maps
per class.

In R_GetTranslation(), the hexen_v10 mode uses slightly different
logic for determining the actual tclass and tmap to use for the mobj.

Todo: Use the same function elsewhere in the code to translate
from the "logical" player class and color to mobj tclass and tmap.