Hexen|Refactor: Unifying logic for setting mobj color translation
It is not a good idea to spread the logic for player color translation
across many places in the code base. Now the final mobj tclass
and tmap are set in a single place, and the player color stored
in the mobj's MF_TRANSLATION is the logical player color.

Todo: Remove the special cases throughout the code base for
swapping tmap colors; it will all be logical colors now.