Renderer: Added a camera vignette effect
In a physical camera lens, less light is received from the edges
of the image compared to the center, making the center appear
brighter. This is called "vignetting" and it can be used to enhance
the appearance of a frame by making it seem more natural and
to subtly help focus the eye on the center of the frame.

See: http://en.wikipedia.org/wiki/Vignetting

The renderer now draws a simple vignette by default. It can be
toggled on and off with "rend-vignette", and further configured
with "rend-vignette-darkness" and "rend-vignette-width".