API|Refactor: Collect map data lumps in WadMapConverter
Locating the map data resources needed for converting a non-native
map format is now the job of the converter plugin.

HOOK_MAP_CONVERT hooks are passed the URI specified to the engine by
the game plugin. The map converter should attempt to resolve said URI
and then locate all associated data lumps needed for the conversion.