Refactor: Moved ddstring, Writer, Reader to libdeng1
This sizable commit moves the Str, Reader and Writer classes
from the engine to libdeng1. "Str" was added as the new
preferred alias for struct ddstring_s, as it is consistent
with the prefix naming convention.

In other words, instead of "ddstring_t*", one should
henceforth use "Str*".

The dd_string.h header was renamed to "de/str.h".

libdeng2: C wrappers were added for the little-endian
byte order conversion functions for various values as
the engine's SHORT(), LONG(), etc. byte order macros are
not available (and shouldn't be) in libdeng1.

The constructors of Reader and Writer that access the
network data buffer were left behind in the engine.

Win32: Obsolete exports were removed from doomsday.def.