Fixed|Game Save|Hexen: Mismanagement of SaveInfo data when saving in Hexen
In Hexen the save state is written to the "base" slot before later
being copied behind the scenes to whichever user-writeable slot is
the destination slot. Therefore we cannot replace the SaveInfo for
the dest slot with that prepared for the base slot as the latter
has been configured with the temporary file paths.

Copying the base slot to the dest slot will automatically result
in updating the relevant info data so simply discard the temporary
SaveInfo in Hexen.

Todo: Unify SaveInfo management across all supported games.