Hacx|libdoom: Fixed bug "[Hacx] Terminatrix not resurrecting foes"
Evidently whatever source port was used during the development and
testing of the Hacx v1.2 IWAD fixes an issue in VileChase action.

In the original id tech 1 version of this routine, the radius used
in the maximum distance is always taken from the MT_VILE mobj type,
regardless of the actual type of the resurrector mobj.

Throw DeHackEd into the equation and the opportunity presents for
the VileChase action to be used with any mobj type, which in the
case of the Hacx Terminatrix, results in a corpse never being near
enough for resurrection.

Added a cvar "game-vilechase-usevileradius" (default: off).
Added a control to the "GameplayOptions" menu page.

See bug: http://sourceforge.net/tracker/index.php?func=detail&aid=3085436&group_id=74815&atid=542099