Deh Reader: Stub for Text patch remapping of Finale backgrounds
Patching Finale backgrounds is not yet supported as presently, there
is no means to route this information through to the Finale script
interpreter.

One solution would be for the games to define a Value definition for
each finale background with an associated name, which could then be
used directly in the InFine script (the script interpreter would then
dynamically lookup the referenced value and substitute the name).