InFine: "flat" commands accept materials specified with a Values URI
The "flat" command in a finale script may now specify the material
to use as the background by declaring a Value definition which holds
the material URI to be resolved.

For example, the command "flat Values:myvalue" will search the DED
database for a Value definition with the id 'myvalue'. If found the
text string associated with that value is interpreted as a material
URI which is then resolved to determine the final material.