Fixed|Bsp Builder: Error loading Doom2 MAP02 (degenerate leaf on back side of line #491)
This special case produces results in one-sided window for one half
the line and a degenerate leaf on the back side of the other. It is
still logical to collapse the degenerate half, so we must therefore
unlink the degenerate from the hedge "on-side" chain as we go.

This map geometry error is now correctly accounted for and there is
no longer any degenerate BSP leafs in the map. Among other things,
this means that it is now possible for a player in the void to move
freely around the map without fear of crashing the game.