Fixed|Bsp Builder: Revised degenerate leaf collapsing - no more orphans
Orphaned leafs are now algorithmically impossible as these are now
collapsed similarly to degenerates in the process of partitioning
the map. Resulting in a reliably correct BSP.

This does not mean that all geometry resulting from bogus mapping
constructs is removed from the BSP. However, this geometry is now
"benign" as it can no longer interfere with the playable areas of
the map.

More work is necessary to fully cleanup from degenerate cases. In
particular the current approach of removing (some) line > sector
references should be replaced with a flag (any such instances are
tagged with $degenleaf in comments).