Optimize|SFX: Further improvements for sector reverb calculations
Only the sector where the listener is in is ever needed for the SFX
playback. It is therefore unnecessary to update the reverb properties
until the listener arrives in a particular sector.

This makes it possible to keep the updates always on even when 3D SFX
is disabled, so that there are no glitches when switching 3D SFX on
and off.