Bsp Builder: Disabled degenerate collapsing
Sadly it appears we will have to continue to put up with orphan leafs
for now. Precision issues in Partitioner result in logically valid
convex polygons being reduced to zero-area ones (e.g., a very thin,
long, triangle where the vertices of the thin edge are so close
together they are considered to exist at the same position in the
plane).

I'll return to this after the Doomsday 1.9.9 release.