libcommon: Experimenting with a console command EventSequence handler
Presently the cheat console commands use a somewhat awkward mechanic
where the arguments are converted to integers, packed into an array
and then fed to the relevant cheat code handler. This not only makes
them difficult to extend but also results in code duplication.

G_AddEventSequenceCommand() works identically to G_AddEventSequence()
except that the native function callback is replaced with a console
command execution.

With a little reqwiring work, most if not all the cheat code handlers
could be replaced with templated console command event sequences.

Presently unused.