Fixed|de::Bsp::Partitioner: Benign logic bug in buildHEdgesAtIntersectionGaps()
When building edges along the hplane partition a bug in the logic for
determining the sector on each side of the partition would result in
changing the "after" sector reference for current HEdgeIntercept.

Benign because the partition hedge builder does not change direction
(and the intercepts cleared after this operation).