Fixed: GL_CalcLuminance() mishandling of all-masked textures
The edge case of an all-masked texture was mishandled resulting in a
brilliant-white origin at the center of the texture. The algorithm
has been revised so that textures with zero color contributing pixels
produce a zero-strength black light origin.

This way, this logic is aligned with that used during luminous source
projection, so that this common DOOM modding idiom does not result in
auto-light sources. (The mod author intends that the object be hidden
completely from view.)