Fixed: Fatal error attempting to play toten.wad
Caused by the use of malformed, expectedly Patch-format sprites which
were not being necessarily checked for when attempting to render the
mobj or when projecting a luminous sources into the map.

Todo: Try to improve Patch validation logic so that this construct is
not warned about in the log(s) (acknowledge it instead).