Fixed|Deh Reader: Tolerate mid-file EOF characters in DeHackEd patches
Seemingly dehacked.exe will ignore EOF characters encountered before
the real end of the file is reached.

A new IgnoreEOF flag was added to DehReader which enables an EOF
tolerant parsing mode (unexpected EOF characters in the patch will
warned about in the log).

This addresses a compatibility issue with various mods such as the
WolfenDoom series.

Also fixed a bug in parser which resulted in it failed to strip any
trailing carriage return characters (benign, due to other handling).