Refactor: Revised interface for R_MiddleMaterialCoversOpening()
The old interface assumed that we would only ever need to determine
this result by deriving all arguments from a linedef. However, this
is not the logic employed by DOOM's software renderer.

The new interface allows for the sector arguments to be specified
directly, so that they may be taken from hedges when appropriate.