Fixed|Map Renderer: Incorrect logic in R_MarkDependantSurfacesForDecorationUpdate()
This function made assumptions about the current relationships of a
linedef's plane interfaces. This routine does not have the necessary
information to determine which of the dependent surfaces are actually
in need of an update and should therefore mark them all.

Logic elsewhere will later determine whether any update work is needed,
such as rebuilding geometry or re-plotting light decorations.

This fixes the bug where light decorations on the middle wall surfaces
of two-sided lines were not always updated following a plane move.