Fixed: HOM before an InFine script at map start
The finale script is not visible until at least one command has been
executed. This is because the script defines the entire appearance of
the screen (background, etc.) and before something is executed we don't
yet know what to draw.

Now one tick is pre-run for all scripts immediately after they've been
loaded.