libcommon: Write a "Saved screenshot" confirmation message to the player's log
Instead of writing the confirmation message to the console, add this
message to the console player's game message log. This should make it
more obvious that a screenshot was actually saved and hopefully avoid
unnecessarily confusing the user.

Todo for later: The games' player message log is not the ideal place
to display such notifications. The engine could feature a notifcation
subsystem that could be leveraged for this sort of thing.