Fixed: Mishandling of malformed Doom-format sound lumps
Some add-ons include malformed Doom-format (i.e., raw PCM) lump data
with the intention being that the sound should not be played.

A basic well-formed test (there is at least one sample and not more
than the number of bytes in the lump) can be used to catch most of
these and suppress loading of the junk data.

See: http://sourceforge.net/p/deng/bugs/1064/