Homepage|Build Repository: Added a releaseDate property to iDownloadable
In the case of symbolic packages, there may be no corresponding
BuildEvent in the build repository and (thus no metadata which
can be queried to derive a release date).

All iDownloadables now have a releaseDate property. When a new
AbstractPackage is derived from a file release of a BuildEvent,
we will set the releaseDate property of the package according
to the startDate property of the build event.

This allows the release date of a downloadable to be assigned
manually if need be.