API: Revised lump cache interface
The older interface was inadequate for a multithreaded architecture
where thread-safe access to the cached data is critical. It was also
error prone and exposed implementation details which users of the
lump cache should not need to know of or be concerned with (that the
cached data is read into Zone-allocated memory).

Now, users of the lump cache no longer specify a Zone purge level as
argument to W_CacheLump() and must now call W_UnlockLump() once done
processing the cached data.