Fixed|Window Manager: Overriding window size with command line options
If the -width and -height (-wh) options are used with the window in
fullscreen mode, the values should not override the normal mode window
geometry.

This addresses an issue with the launcher where the stored normal-mode
size was always lost, as the window dimensions are always defined on
the command line.