Refactor: Use a DFile in AudioDriver musicPlayLump()
Instead of reading directly from the lump using F_ReadLumpSection()
when buffering the song, instead open a new DFile handle on it and
read with that.

DFile completely abstracts the underlying implementation such that
any file, whether it be a native system file or a lump in a container
(such as WadFile) can be manipulated indirectly with one interface.

Todo: Reimplement all users of the lower level F_Read* family of
functions outside of the file system layer to use DFile.