Fixed: Many C++ namespace problems in recently refactored code
The core issue here is that the global namespace is polluted by both
opaque handle typedefs and similarly named classes brought in from the
de namespace. This practice should be discontinued. The opaque handles
should only be visible in C code and in the code that defines the
C wrapper functions -- never in C++ code where the actual classes are
available.

Also, instead of "using", putting the class method definitions inside
the namespace will tell the compiler to prioritize that namespace over
the global one, resolving some of the identifier lookups.