Refactor|FileSys: Removed redundant (duplicated) file system search algorithms
FS::collectFiles() has been replaced with FS::findAll(), which returns
a FileList instead of an array of AbstractFile pointers.

FS::collectPaths() has been replaced with a variant of FS::findAll()
that takes a "match predicate" function pointer which is called to
decide if a file is included in the result.

FS::collectLocalPaths() is no longer a public API function and is
now used internally by the FS class only.