Refactor|FileSys: Ordain WadFile lumps named "DEHACKED" a ".deh" extension
Now that all files support an extension including those loaded from
WAD files, it is no longer necessary for special case kludges so that
these are picked up by the DehRead plugin.

Simply ordain any lump named "DEHACKED" with a ".deh" extension when
reading the lump directory in WadFile and it will be detected in the
same way as real files.